Consolidate spec files
