+++ spec/rails_spec.rb
@@ -52,3 +52,31 @@ RSpec.describe "PostgreSQLGuard" do
   end
 
 end
+
+
+RSpec.describe "MethodCollision" do
+  it "including model does not override an existing search method" do
+    klass = Class.new(ActiveRecord::Base) do
+      self.table_name = :mock_items
+
+      def self.search(_query)
+        :custom_search
+      end
+
+      include ParadeDB::Model
+    end
+
+    expect(klass.search("query")).to eq(:custom_search)
+    expect(klass).to respond_to(:paradedb_search)
+  end
+
+  it "including model does not raise without collision" do
+    klass = Class.new(ActiveRecord::Base) do
+      self.table_name = :mock_items
+      include ParadeDB::Model
+    end
+
+    expect(klass).to respond_to(:search)
+    expect(klass).to respond_to(:paradedb_search)
+  end
+end
